Metric based registration marks searching.
	modified:   src/fresco/CImageContainer.java
	modified:   src/fresco/swing/CImagePanel.java
	modified:   src/geneticalgorithms/CGenotype.java
	modified:   src/geneticalgorithms/CRouletteWheel.java
	modified:   src/image/CBinaryImage.java
	modified:   src/image/colour/CBilinearInterpolation.java
	deleted:    src/image/converters/Crgb2gray.java
	new file:   src/image/converters/Crgb2grey.java
	modified:   src/image/statiscics/CHistogram.java
	new file:   src/utils/ArrayUtils.java
	modified:   src/utils/geometry/CPerspectiveTransformation.java
	new file:   src/utils/geometry/CPointAndTransformation.java
	new file:   src/utils/metrics/CAreaSimilarityMetric.java
	new file:   src/utils/metrics/CCovarianceMetric.java
	new file:   src/utils/metrics/CCrossCorrelationMetric.java
	modified:   src/utils/metrics/CEuclidMetrics.java
	modified:   src/utils/vector/CTreshold.java
	modified:   src/workers/analyse/CDiff.java
	modified:   src/workers/analyse/CPatternAnalyzer.java
	modified:   src/workers/analyse/CVariationOfInformationGraph.java
	modified:   src/workers/registration/CPerspectiveTransformationWorker.java
	modified:   src/workers/registration/CPointPairs.java
	modified:   src/workers/registration/refpointga/CCrossCorrelationFitness.java
	modified:   src/workers/registration/refpointga/CNearestEdgeMatrix.java
	modified:   src/workers/registration/refpointga/CPointAndTransformation.java
	modified:   src/workers/registration/refpointga/CPointMutation.java
	modified:   src/workers/registration/refpointga/CRefPointMarker.java
	modified:   test/image/converters/Crgb2grayTest.java
	modified:   test/workers/registration/refpointga/CNearestEdgeMatrixTest.java
	modified:   test/workers/registration/refpointga/CNearestPointCompetitorTest.java

+++ src/workers/registration/refpointga/CPointMutation.java
@@ -22,7 +22,7 @@ class CPointMutation implements IMutationOperator<CPointAndTransformation> {
 	/** gimli based mutation */
 	public void mutateByGimli(CPointAndTransformation prototype) {
 
-		int x, y;
+		double x, y;
 		double dx, dy, angle, scale;
 
 		if (Math.random() < MUTATION_PROBABBILITY) {
@@ -59,7 +59,7 @@ class CPointMutation implements IMutationOperator<CPointAndTransformation> {
 
 	public void mutate(CPointAndTransformation prototype) {
 
-		int x, y;
+		double x, y;
 		double dx, dy, angle, scale;
 
 		double rangeFactor = CRefPointMarker.MUTATION_FACTOR * (1 - prototype.getFitness());
